+++ pkg/backup/operator/event/event.go
@@ -26,7 +26,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/rs/zerolog/log"
+	"github.com/rs/zerolog"
+
 	core "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/uuid"
@@ -34,10 +35,11 @@ import (
 )
 
 // NewEventRecorder creates new event recorder
-func NewEventRecorder(name string, kubeClientSet kubernetes.Interface) Recorder {
+func NewEventRecorder(logger zerolog.Logger, name string, kubeClientSet kubernetes.Interface) Recorder {
 	return &eventRecorder{
 		kubeClientSet: kubeClientSet,
 		name:          name,
+		logger:        logger,
 	}
 }
 
@@ -51,6 +53,7 @@ type Recorder interface {
 type eventRecorder struct {
 	name          string
 	kubeClientSet kubernetes.Interface
+	logger        zerolog.Logger
 }
 
 func (e *eventRecorder) newEvent(group, version, kind string, object meta.Object, eventType, reason, message string) *core.Event {
@@ -90,7 +93,7 @@ func (e *eventRecorder) newObjectReference(group, version, kind string, object m
 func (e *eventRecorder) event(group, version, kind string, object meta.Object, eventType, reason, message string) {
 	_, err := e.kubeClientSet.CoreV1().Events(object.GetNamespace()).Create(context.Background(), e.newEvent(group, version, kind, object, eventType, reason, message), meta.CreateOptions{})
 	if err != nil {
-		log.Warn().Err(err).
+		e.logger.Warn().Err(err).
 			Str("APIVersion", fmt.Sprintf("%s/%s", group, version)).
 			Str("Kind", kind).
 			Str("Object", fmt.Sprintf("%s/%s", object.GetNamespace(), object.GetName())).
@@ -98,7 +101,7 @@ func (e *eventRecorder) event(group, version, kind string, object meta.Object, e
 		return
 	}
 
-	log.Info().
+	e.logger.Info().
 		Str("APIVersion", fmt.Sprintf("%s/%s", group, version)).
 		Str("Kind", kind).
 		Str("Object", fmt.Sprintf("%s/%s", object.GetNamespace(), object.GetName())).