+++ pkg/logging/const.go
@@ -1,0 +1,49 @@
+//
+// DISCLAIMER
+//
+// Copyright 2021 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+// Author Adam Janikowski
+//
+
+package logging
+
+const (
+	LoggerNameOperator              = "operator"
+	LoggerNameDeployment            = "deployment"
+	LoggerNameKLog                  = "klog"
+	LoggerNameServer                = "server"
+	LoggerNameDeploymentReplication = "deployment-replication"
+	LoggerNameStorage               = "storage"
+	LoggerNameProvisioner           = "provisioner"
+	LoggerNameReconciliation        = "reconciliation"
+	LoggerNameEventRecorder         = "event-recorder"
+)
+
+func LoggerNames() []string {
+	return []string{
+		LoggerNameOperator,
+		LoggerNameDeployment,
+		LoggerNameKLog,
+		LoggerNameServer,
+		LoggerNameDeploymentReplication,
+		LoggerNameStorage,
+		LoggerNameProvisioner,
+		LoggerNameReconciliation,
+		LoggerNameEventRecorder,
+	}
+}